+++ src/stories/BannerHeader.stories.js
@@ -17,13 +17,8 @@ const mock = {
     category: "Lectus",
     startDate: "1995-12-17T03:24:00",
     endDate: "1995-12-17T03:24:00",
-    byline: [
-        {
-            id: "3062",
-            title: "Jen Diamond",
-            to: "",
-        },
-    ],
+    byline: ["Jen Diamond", "Pikku Lehtonen"],
+    subjectAreas: [{ title: "Ghosts" }, { title: "Clowns" }],
     locations: [
         {
             id: "523",
@@ -105,7 +100,7 @@ export const NoCategory = () => ({
     `,
 })
 
-export const WithByline = () => ({
+export const WithBylineAndSubjectAreas = () => ({
     data() {
         return {
             ...mock,
@@ -120,6 +115,7 @@ export const WithByline = () => ({
            :start-date="startDate"
            :end-date="endDate"
            :byline="byline"
+           :subjectAreas="subjectAreas"
            :prompt="prompt"
            :align-right="false"
        />
@@ -189,14 +185,8 @@ const article = {
     title: "I Like Turtles",
     category: "Library News",
     dateCreated: "2022-02-09T10:57:46-08:00",
-    byline: [
-        {
-            title: " Written by Courtney Hoffner",
-        },
-        {
-            title: "Illustrations by Jen Diamond",
-        },
-    ],
+    byline: [" Written by Courtney Hoffner", "Illustrations by Jen Diamond"],
+    subjectAreas: [{ title: "Ghosts" }, { title: "Clowns" }],
     locations: [
         {
             id: "523",
@@ -225,6 +215,7 @@ export const ArticleDetail = () => ({
            :image="image"
            :title="title"
            :byline="byline"
+           :subjectAreas="subjectAreas"
            :dateCreated="dateCreated"
            :locations="locations"
            :text="text"