+++ src/stories/BlockHighlight.stories.js
@@ -37,7 +37,7 @@ const mock = {
 }
 
 const mock2 = {
-    image: API.image,
+    image: null, //API.image,
     to: "/about/foo/bar/",
     category: "Lerisque",
     title: "Orci Dapibus",