+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.68.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.68.0...v1.68.1) (2022-11-08)
+
+
+### Bug Fixes
+
+* Impact Report home page cannot be a nuxt link on drupal environment ([#234](https://github.com/UCLALibrary/ucla-library-website-components/issues/234)) ([ba5efcd](https://github.com/UCLALibrary/ucla-library-website-components/commit/ba5efcd1086c1a002e18107e1878be66208737b0))
+
+# [1.68.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.67.1...v1.68.0) (2022-11-07)
+
+
+### Features
+
+* Add subject area to banner text and header ([#233](https://github.com/UCLALibrary/ucla-library-website-components/issues/233)) ([479f458](https://github.com/UCLALibrary/ucla-library-website-components/commit/479f45888e3c1c1b4f181712e464af6c338f6b4d))
+
+## [1.67.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.67.0...v1.67.1) (2022-11-07)
+
+
+### Bug Fixes
+
+* add molecule svg with blue background when there is no event image prop for blockhighlight ([#232](https://github.com/UCLALibrary/ucla-library-website-components/issues/232)) ([5261c78](https://github.com/UCLALibrary/ucla-library-website-components/commit/5261c78ed5d28e0d1c3a2c614bb6c658ecf6fd06))
+
 # [1.67.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v1.66.1...v1.67.0) (2022-11-02)